airflow/requirements.txt
requests
python-dotenv
snowflake-sqlalchemy
bs4
datetime
apache-airflow==2.8.1